+++ CHANGELOG.md
@@ -125,6 +125,7 @@
 - Add recycle bin endpoints @robgietema
 - Change format of system endpoint @robgietema
 - Profiles are now specified by package name and profile name @robgietema
+- Check accept header on returning content @robgietema
 
 ### Bugfix